Added virtual environment + made changes to main code file
A lot to go through here but I'll keep it brief. I added in a virtual environment to make working with the required modules easier. It also allows the program to be used easier on other systems. I adjusted the main program's comments and added some global variables. I added a couple things that I will be using for developing a GUI which right now is just the file KeystrokeApp.py and folder templates.