Added in basic PyQt5 window and made minor adjustments to KeystrokeBase.py
I have decided to use PyQt5 as the main GUI framework. It's a lot more simple and easier to implement as it is implemented purely with Python as compared to Eel and Electron which use web styling. In the future we may switch to a web style but for now, I am going to stick with PyQt5. Requirements.txt has been updated accordingly.

KeystrokeBase.py
- Standardized error messages
- Simplified some comments
- Basic style changes
- Added in loadJSON function to load in a JSON and its info
- Small bug fixes
- Changed file open to more standard method (using with open())
- Changed keystroke print to be within getKeystrokes (to maintain proper scoping)

app.py:
- Added in basic window without any functionality
- Successfully added in PyQt5 library